+++ src/adc/writer/speech/_txt.py
@@ -12,6 +12,7 @@ from adc.api import SpeechData
 class TxtSpeechWriter(SplittableStreamWriter, AnnotationsOnlyWriter, InputBasedPlaceholderSupporter):
 
     def __init__(self, output_dir: str = None, annotations_only: bool = None,
+                 speaker_suffix: str = None, speaker_key: str = None,
                  split_names: List[str] = None, split_ratios: List[int] = None, split_group: str = None,
                  logger_name: str = None, logging_level: str = LOGGING_WARNING):
         """
@@ -21,6 +22,10 @@ class TxtSpeechWriter(SplittableStreamWriter, AnnotationsOnlyWriter, InputBasedP
         :type output_dir: str
         :param annotations_only: whether to output only the annotations and not the images
         :type annotations_only: bool
+        :param speaker_suffix: the file suffix for the files containing the speaker name (incl dot)
+        :type speaker_suffix: str
+        :param speaker_key: the key for the speaker name/id in the meta-data
+        :type speaker_key: str
         :param split_names: the names of the splits, no splitting if None
         :type split_names: list
         :param split_ratios: the integer ratios of the splits (must sum up to 100)
@@ -35,6 +40,8 @@ class TxtSpeechWriter(SplittableStreamWriter, AnnotationsOnlyWriter, InputBasedP
         super().__init__(split_names=split_names, split_ratios=split_ratios, split_group=split_group, logger_name=logger_name, logging_level=logging_level)
         self.output_dir = output_dir
         self.annotations_only = annotations_only
+        self.speaker_suffix = speaker_suffix
+        self.speaker_key = speaker_key
 
     def name(self) -> str:
         """
@@ -52,7 +59,7 @@ class TxtSpeechWriter(SplittableStreamWriter, AnnotationsOnlyWriter, InputBasedP
         :return: the description
         :rtype: str
         """
-        return "Saves the transcript in a .txt file alongside the audio file."
+        return "Saves the transcript in a .txt file alongside the audio file. If a speaker suffix is defined, then available speaker information gets stored in an additional companion file."
 
     def _create_argparser(self) -> argparse.ArgumentParser:
         """
@@ -64,6 +71,8 @@ class TxtSpeechWriter(SplittableStreamWriter, AnnotationsOnlyWriter, InputBasedP
         parser = super()._create_argparser()
         parser.add_argument("-o", "--output", type=str, help="The directory to store the audio/.txt files in. Any defined splits get added beneath there. " + placeholder_list(obj=self), required=True)
         add_annotations_only_writer_param(parser)
+        parser.add_argument("--speaker_suffix", type=str, help="The file suffix for the companion files to store the speaker in, e.g., '.speaker'.", required=False, default=None)
+        parser.add_argument("--speaker_key", type=str, help="The key in the meta-data with the speaker name/ID.", required=False, default="speaker")
         return parser
 
     def _apply_args(self, ns: argparse.Namespace):
@@ -76,6 +85,8 @@ class TxtSpeechWriter(SplittableStreamWriter, AnnotationsOnlyWriter, InputBasedP
         super()._apply_args(ns)
         self.output_dir = ns.output
         self.annotations_only = ns.annotations_only
+        self.speaker_suffix = ns.speaker_suffix
+        self.speaker_key = ns.speaker_key
 
     def accepts(self) -> List:
         """
@@ -93,6 +104,8 @@ class TxtSpeechWriter(SplittableStreamWriter, AnnotationsOnlyWriter, InputBasedP
         super().initialize()
         if self.annotations_only is None:
             self.annotations_only = False
+        if self.speaker_key is None:
+            self.speaker_key = "speaker"
 
     def write_stream(self, data):
         """
@@ -120,3 +133,13 @@ class TxtSpeechWriter(SplittableStreamWriter, AnnotationsOnlyWriter, InputBasedP
                 with open(path, "w") as fp:
                     fp.write(item.annotation)
                     fp.write("\n")
+
+            if self.speaker_suffix is not None:
+                if item.has_metadata() and (self.speaker_key in item.get_metadata()):
+                    path = os.path.splitext(path)[0] + self.speaker_suffix
+                    self.logger().info("Writing speaker to: %s" % path)
+                    with open(path, "w") as fp:
+                        fp.write(item.get_metadata()[self.speaker_key])
+                        fp.write("\n")
+                else:
+                    self.logger().warning("No speaker available from meta-data: %s" % item.audio_name)